updated feature file and new set up system
for the new set up you will have to put your info into the Enviornment set up file and then move that to a folder that is not shared on github. before running these files you will have to run the set up file.
for the features doc I added aac and worked on only having one speed with limited success